[#549] fix Thrift server AWS auth locally
In #502, we made the decision to push IRSA-based authentication as the
default. I think this is still worthwhile, but the shift has to be more
holistic.  There are still several other services that assume the AWS
enviroment variables will be provided via the remote-auth-config secret.
To really change the default to IRSA, we likely want to flow a separate
ticket and fully think through the best approach, and unify the auth
method across all deployments.

For now, we will simply keep the secret-based auth as the default.