[#585] fix Nvidia ARM build
This change set does two things:
 - Add a kustomization to fix the Docker `dind` image to a specific
   version (v27.5.1)
 - Update the buildkit image tag for the docker-maven-plugin to a fixed
   version (v0.19.0)

The original issue appeared like it might be caused to the updated
`dind` image which corresponded with the release of Docker Engine 28.0.
The actual cause ended up being the update to the `buildkit-stable-1`
tag of the buildkit image we were using (also corresponding to the
engine).