[#591] fix tty error in CI
The CI runner has no TTY session. Instead of tee-ing the output to
`/dev/tty` and adding processing logic further along the pipe, we have
to use a tee'd subshell for additional processing and allow the main
output to continue unprocessed to stdout. This means we can't do
straight variable capture so a file is used to hold the processed output
instead.