[#622] fix default configmap value for hive chart
The property `configMap.metastoreServiceConfig.properties` is intended
to be a list of objects representing additional configs'
name/value/descriptions. However, the default value in `values.yaml`
was an empty object/map causing a warning when the chart was used as
intended: _cannot overwrite table with non table_.

Unfortunately, the `helm-unittest` plugin does not pick up this scenario,
so our automated tests did not catch this and it is unclear why.