[#642] various build optimizations
Docker Improvements:
Moves the provisioning and teardown of the k8s docker builder to the GH
action to avoid the Maven build cache skipping the logic. Reduces the
cache level to the default `min` instead of max, as it _appears_ that
the layer cache isn't really working with the k8s builder, just the base
image cache. Skips the image build in CI by default (like the previous
behavior) with images that have tests specifically turning the image
build back on to ensure the tests are run against the latest image.

Archetype improvements:
Uses generate-sources for final check in archetype test. We still do a
final build to verify the project builds successfully, but the check for
manual actions is done separately to avoid a full rebuild. This should
save about 9-10 minutes.

GH Cache Improvments:
Uses the `github-script` action to call the same hashing function used
by the `runs-on/cache` action to detect whether caches actually had
their contents modified, regardless of cache-hit. This saves time when
some POM or Dockerfile change did not actually result in new
dependencies, or new layers. Because of the inherent limitations of the
way the gh runner logic parses
[`hashFiles`](https://github.com/actions/runner/blob/c1095ae2d100886a7865f466357f09ef30dec1b9/src/Runner.Worker/Expressions/HashFilesFunction.cs#L44)
we aren't able to directly call `hashFiles` with a [different working
directory](https://github.com/actions/toolkit/issues/1035#issuecomment-1678272606)
so instead, we use the `github-script` action to call `glob.hashFiles`
directly.