[#639] cleanup/fix dependencies in extensions-transform-spark-java
Clean up dependencies in extensions-transform, and update the Sedona
library to the latest version so we can grab a Spark 3.5 compatible
version of the jars. See the [Sedona docs](https://sedona.apache.org/1.7.1/setup/maven-coordinates/)
for details. This wasn't an issue before just by random chance, but the
older version of Sedona causes a conflict with the commons-text version
on the classpath.