[#670] makes drift-detection client ITs build-cache compatible
We were using a relative path to the drift detection service build
output directory to start the quarkus-run jar for the ITs. This only
works if the quarkus-app directory is present, which it is not if the
build cache causes the service module to be skipped.

This uses the same Quarkus app pattern we've used for our dockerized
Quarkus apps in order to ensure the fully-packaged app is available as
an artifact, and thus restored by the build-cache and retrievable from
the .m2 cache.

This became an issue after the removal of `quarkus-app` from the
attached outputs in #665