[#690] updates MLFlow to 2.10.2
We did not update to the latest version due to a bug with the 1.8
version of poetry-plugin-export. Also bumped the minimum Python up to
3.9 since 2.8 was EOL in October 2024.

Fixes a bug with the MDA pyproject.toml templates that was causing only
generated code to be included in a pipeline instead of the full source.