create the builders in fabric8s config directory
We can't directly use `--config ~/.docker/fabric8` because it will not
be present with the buildx plugin on a fresh action runner. Instead, we
just wholesale copy the setup from `.docker` into the fabric8 directory.
This allows for the credentials to be preserved in the root `.docker`
directory between image builds, and still allows us to do default
builder management in the build setup.

Also groups the install dependencies steps for more readable output.