[#720] use python action instead of pyenv
Mostly this just involves updating the `install_dependencies` action and
updating Habushu's `usePyenv` to `false` if the `ci` profile is active.
However, because the archetype test _also_ builds Habushu modules, it
was trying to use `pyenv`.  So the foundation-archetype pom/test was
updated to allow passing build profiles to the test script, and if the
`ci` profile is active the module will pass `-Pci` to the script.

Also fixes the `RuffTomlFileGenerationMigration` test. Apparently MacOs
treats file paths case-insensitive (at least when accessed via Java)
whereas the Ubuntu image we're using in CI does not.