[#716] update MLFlow to 2.22.0
Techinically 2.21.2 would fix all of the vulnerabilities, but I figured
we might as well use the latest as the minimum. Because of transitive
dependency incompatibilities, had to also allow Pandas to upgrade to 2+
and remove the scipy set version of 1.9.3

Also bumped Python minimum up to 3.9 in a few places that were missed in
the previous bump.