[#733] incorporate dependapot suggestions
Primarly just touches Maven plugins like the archetype. Because we've
add Habushu as a dependency in order to help with build cache
calculations, a transitive dependency there was making dependabot think
we were using a vulnerable version of a library in our application code.
I added exclusions to the dependencies we added for this purpose just to
reduce the noise.

This exposed a few places where we were actually relying on the
transitive dependencies brought in by those for, namely, Krausening and
aether-api.